Move batcheshelper out of enterprise (#54923)
Not required anymore, just a git mv and some path changes.

CI.

(cherry picked from commit 9a3715063e7bfb7c85e6b7b82cd0b53221003b12)
