+++ cmd/batcheshelper/image_test.yaml
