+++ cmd/batcheshelper/log/BUILD.bazel
@@ -2,25 +2,22 @@ load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "util",
-    srcs = [
-        "exec.go",
-        "util.go",
-    ],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/batcheshelper/util",
+    name = "log",
+    srcs = ["log.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/batcheshelper/log",
     visibility = ["//visibility:public"],
     deps = [
-        "//internal/executor/types",
-        "//internal/executor/util",
+        "//lib/batches",
         "//lib/errors",
     ],
 )
 
 go_test(
-    name = "util_test",
-    srcs = ["util_test.go"],
+    name = "log_test",
+    srcs = ["log_test.go"],
     deps = [
-        ":util",
+        ":log",
+        "//lib/batches",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
     ],
