+++ cmd/batcheshelper/run/BUILD.bazel
@@ -7,11 +7,11 @@ go_library(
         "post.go",
         "pre.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/batcheshelper/run",
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/batcheshelper/run",
     visibility = ["//visibility:public"],
     deps = [
-        "//enterprise/cmd/batcheshelper/log",
-        "//enterprise/cmd/batcheshelper/util",
+        "//cmd/batcheshelper/log",
+        "//cmd/batcheshelper/util",
         "//internal/executor/types",
         "//lib/batches",
         "//lib/batches/execution",
@@ -31,8 +31,8 @@ go_test(
     ],
     deps = [
         ":run",
-        "//enterprise/cmd/batcheshelper/log",
-        "//enterprise/cmd/batcheshelper/util",
+        "//cmd/batcheshelper/log",
+        "//cmd/batcheshelper/util",
         "//lib/batches",
         "//lib/batches/env",
         "//lib/batches/execution",
