+++ cmd/batcheshelper/run/pre.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/kballard/go-shellquote"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/batcheshelper/log"
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/batcheshelper/util"
+	"github.com/sourcegraph/sourcegraph/cmd/batcheshelper/log"
+	"github.com/sourcegraph/sourcegraph/cmd/batcheshelper/util"
 	"github.com/sourcegraph/sourcegraph/internal/executor/types"
 	batcheslib "github.com/sourcegraph/sourcegraph/lib/batches"
 	"github.com/sourcegraph/sourcegraph/lib/batches/execution"
