+++ cmd/batcheshelper/util/BUILD.bazel
@@ -2,22 +2,25 @@ load("//dev:go_defs.bzl", "go_test")
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "log",
-    srcs = ["log.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/batcheshelper/log",
+    name = "util",
+    srcs = [
+        "exec.go",
+        "util.go",
+    ],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/batcheshelper/util",
     visibility = ["//visibility:public"],
     deps = [
-        "//lib/batches",
+        "//internal/executor/types",
+        "//internal/executor/util",
         "//lib/errors",
     ],
 )
 
 go_test(
-    name = "log_test",
-    srcs = ["log_test.go"],
+    name = "util_test",
+    srcs = ["util_test.go"],
     deps = [
-        ":log",
-        "//lib/batches",
+        ":util",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
     ],
