+++ cmd/batcheshelper/util/util.go
