+++ enterprise/cmd/bundled-executor/BUILD.bazel
@@ -5,7 +5,7 @@ load("//dev:oci_defs.bzl", "image_repository")
 load("//cmd/server:macro.bzl", "container_dependencies", "dependencies_tars")
 
 DEPS = [
-    "//enterprise/cmd/batcheshelper",
+    "//cmd/batcheshelper",
     "//enterprise/cmd/executor",
 ]
